fix: regenerate lockfile for ast-parser workspace package
